backend/requirements.txt
fastapi>=0.114.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
motor>=3.5.0
beanie>=1.26.0
pydantic>=2.7.0
